+++ src/main/java/com/alternacraft/pvptitles/Hook/PlaceholderHook.java
@@ -52,7 +52,7 @@ public class PlaceholderHook extends EZPlaceholderHook {
             LoggerManager.logError(ex.getCustomMessage());
         }
 
-        int seconds = 0;
+        long seconds = 0;
         try {
             seconds = plugin.getManager().dbh.getDm().loadPlayedTime(player.getUniqueId());
         } catch (DBException ex) {
